+++ src/components/input/input.test.ts
@@ -85,38 +85,73 @@ describe('<sl-input>', () => {
   it('should focus the input when clicking on the label', async () => {
     const el = await fixture<SlInput>(html` <sl-input label="Name"></sl-input> `);
     const label = el.shadowRoot!.querySelector('[part~="form-control-label"]')!;
-    const submitHandler = sinon.spy();
+    const focusHandler = sinon.spy();
 
-    el.addEventListener('sl-focus', submitHandler);
+    el.addEventListener('sl-focus', focusHandler);
     (label as HTMLLabelElement).click();
-    await waitUntil(() => submitHandler.calledOnce);
+    await waitUntil(() => focusHandler.calledOnce);
 
-    expect(submitHandler).to.have.been.calledOnce;
+    expect(focusHandler).to.have.been.calledOnce;
   });
 
   describe('when using constraint validation', () => {
     it('should be valid by default', async () => {
       const el = await fixture<SlInput>(html` <sl-input></sl-input> `);
-      expect(el.invalid).to.be.false;
+      expect(el.checkValidity()).to.be.true;
     });
 
     it('should be invalid when required and empty', async () => {
       const el = await fixture<SlInput>(html` <sl-input required></sl-input> `);
       expect(el.reportValidity()).to.be.false;
-      expect(el.invalid).to.be.true;
-    });
-
-    it('should be invalid when the pattern does not match', async () => {
-      const el = await fixture<SlInput>(html` <sl-input pattern="^test" value="fail"></sl-input> `);
-      expect(el.invalid).to.be.true;
-      expect(el.reportValidity()).to.be.false;
+      expect(el.checkValidity()).to.be.false;
     });
 
     it('should be invalid when required and disabled is removed', async () => {
       const el = await fixture<SlInput>(html` <sl-input disabled required></sl-input> `);
       el.disabled = false;
       await el.updateComplete;
-      expect(el.invalid).to.be.true;
+      expect(el.checkValidity()).to.be.false;
+    });
+
+    it('should receive the correct validation attributes ("states") when valid', async () => {
+      const el = await fixture<SlInput>(html` <sl-input required value="a"></sl-input> `);
+
+      expect(el.checkValidity()).to.be.true;
+      expect(el.hasAttribute('data-required')).to.be.true;
+      expect(el.hasAttribute('data-optional')).to.be.false;
+      expect(el.hasAttribute('data-invalid')).to.be.false;
+      expect(el.hasAttribute('data-valid')).to.be.true;
+      expect(el.hasAttribute('data-user-invalid')).to.be.false;
+      expect(el.hasAttribute('data-user-valid')).to.be.false;
+
+      el.focus();
+      await el.updateComplete;
+      await sendKeys({ press: 'b' });
+      await el.updateComplete;
+
+      expect(el.checkValidity()).to.be.true;
+      expect(el.hasAttribute('data-user-invalid')).to.be.false;
+      expect(el.hasAttribute('data-user-valid')).to.be.true;
+    });
+
+    it('should receive the correct validation attributes ("states") when invalid', async () => {
+      const el = await fixture<SlInput>(html` <sl-input required></sl-input> `);
+
+      expect(el.hasAttribute('data-required')).to.be.true;
+      expect(el.hasAttribute('data-optional')).to.be.false;
+      expect(el.hasAttribute('data-invalid')).to.be.true;
+      expect(el.hasAttribute('data-valid')).to.be.false;
+      expect(el.hasAttribute('data-user-invalid')).to.be.false;
+      expect(el.hasAttribute('data-user-valid')).to.be.false;
+
+      el.focus();
+      await el.updateComplete;
+      await sendKeys({ press: 'a' });
+      await sendKeys({ press: 'Backspace' });
+      await el.updateComplete;
+
+      expect(el.hasAttribute('data-user-invalid')).to.be.true;
+      expect(el.hasAttribute('data-user-valid')).to.be.false;
     });
   });
 
@@ -166,6 +201,26 @@ describe('<sl-input>', () => {
 
       expect(keydownHandler).to.have.been.calledOnce;
       expect(submitHandler).to.not.have.been.called;
+    });
+
+    it('should be invalid when setCustomValidity() is called with a non-empty value', async () => {
+      const input = await fixture<HTMLFormElement>(html` <sl-input></sl-input> `);
+
+      input.setCustomValidity('Invalid selection');
+      await input.updateComplete;
+
+      expect(input.checkValidity()).to.be.false;
+      expect(input.hasAttribute('data-invalid')).to.be.true;
+      expect(input.hasAttribute('data-valid')).to.be.false;
+      expect(input.hasAttribute('data-user-invalid')).to.be.false;
+      expect(input.hasAttribute('data-user-valid')).to.be.false;
+
+      input.focus();
+      await sendKeys({ type: 'test' });
+      await input.updateComplete;
+
+      expect(input.hasAttribute('data-user-invalid')).to.be.true;
+      expect(input.hasAttribute('data-user-valid')).to.be.false;
     });
   });
 
@@ -261,7 +316,7 @@ describe('<sl-input>', () => {
       await el.updateComplete;
     });
 
-    it('should not emit sl-change or sl-input when calling setRangeText()', async () => {
+    it('should not emit sl-change or sl-input when calling setinputText()', async () => {
       const el = await fixture<SlInput>(html` <sl-input value="hi there"></sl-input> `);
 
       el.addEventListener('sl-change', () => expect.fail('sl-change should not be emitted'));
@@ -277,21 +332,21 @@ describe('<sl-input>', () => {
   describe('when type="number"', () => {
     it('should be valid when the value is within the boundary of a step', async () => {
       const el = await fixture<SlInput>(html` <sl-input type="number" step=".5" value="1.5"></sl-input> `);
-      expect(el.invalid).to.be.false;
+      expect(el.checkValidity()).to.be.true;
     });
 
     it('should be invalid when the value is not within the boundary of a step', async () => {
       const el = await fixture<SlInput>(html` <sl-input type="number" step=".5" value="1.25"></sl-input> `);
-      expect(el.invalid).to.be.true;
+      expect(el.checkValidity()).to.be.false;
     });
 
     it('should update validity when step changes', async () => {
       const el = await fixture<SlInput>(html` <sl-input type="number" step=".5" value="1.5"></sl-input> `);
-      expect(el.invalid).to.be.false;
+      expect(el.checkValidity()).to.be.true;
 
       el.step = 1;
       await el.updateComplete;
-      expect(el.invalid).to.be.true;
+      expect(el.checkValidity()).to.be.false;
     });
 
     it('should increment by step when stepUp() is called', async () => {