+++ src/components/input/input.ts
@@ -4,7 +4,7 @@ import { classMap } from 'lit/directives/class-map.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { live } from 'lit/directives/live.js';
 import { defaultValue } from '../../internal/default-value';
-import { FormSubmitController } from '../../internal/form';
+import { FormControlController } from '../../internal/form';
 import ShoelaceElement from '../../internal/shoelace-element';
 import { HasSlotController } from '../../internal/slot';
 import { watch } from '../../internal/watch';
@@ -63,14 +63,13 @@ const isFirefox = isChromium ? false : navigator.userAgent.includes('Firefox');
 export default class SlInput extends ShoelaceElement implements ShoelaceFormControl {
   static styles: CSSResultGroup = styles;
 
-  private readonly formSubmitController = new FormSubmitController(this);
+  private readonly formControlController = new FormControlController(this);
   private readonly hasSlotController = new HasSlotController(this, 'help-text', 'label');
   private readonly localize = new LocalizeController(this);
 
   @query('.input__control') input: HTMLInputElement;
 
   @state() private hasFocus = false;
-  @state() invalid = false;
   @property() title = ''; // make reactive to pass through
 
   /**
@@ -220,7 +219,7 @@ export default class SlInput extends ShoelaceElement implements ShoelaceFormCont
   }
 
   firstUpdated() {
-    this.invalid = !this.checkValidity();
+    this.formControlController.updateValidity();
   }
 
   private handleBlur() {
@@ -250,12 +249,12 @@ export default class SlInput extends ShoelaceElement implements ShoelaceFormCont
 
   private handleInput() {
     this.value = this.input.value;
-    this.invalid = !this.checkValidity();
+    this.formControlController.updateValidity();
     this.emit('sl-input');
   }
 
   private handleInvalid() {
-    this.invalid = true;
+    this.formControlController.setValidity(false);
   }
 
   private handleKeyDown(event: KeyboardEvent) {
@@ -272,7 +271,7 @@ export default class SlInput extends ShoelaceElement implements ShoelaceFormCont
         // See https://github.com/shoelace-style/shoelace/pull/988
         //
         if (!event.defaultPrevented && !event.isComposing) {
-          this.formSubmitController.submit();
+          this.formControlController.submit();
         }
       });
     }
@@ -284,9 +283,8 @@ export default class SlInput extends ShoelaceElement implements ShoelaceFormCont
 
   @watch('disabled', { waitUntilFirstUpdate: true })
   handleDisabledChange() {
-    // Disabled form controls are always valid, so we need to recheck validity when the state changes
-    this.input.disabled = this.disabled;
-    this.invalid = !this.checkValidity();
+    // Disabled form controls are always valid
+    this.formControlController.setValidity(this.disabled);
   }
 
   @watch('step', { waitUntilFirstUpdate: true })
@@ -294,13 +292,13 @@ export default class SlInput extends ShoelaceElement implements ShoelaceFormCont
     // If step changes, the value may become invalid so we need to recheck after the update. We set the new step
     // imperatively so we don't have to wait for the next render to report the updated validity.
     this.input.step = String(this.step);
-    this.invalid = !this.checkValidity();
+    this.formControlController.updateValidity();
   }
 
   @watch('value', { waitUntilFirstUpdate: true })
   async handleValueChange() {
     await this.updateComplete;
-    this.invalid = !this.checkValidity();
+    this.formControlController.updateValidity();
   }
 
   /** Sets focus on the input. */
@@ -375,10 +373,10 @@ export default class SlInput extends ShoelaceElement implements ShoelaceFormCont
     return this.input.reportValidity();
   }
 
-  /** Sets a custom validation message. If `message` is not empty, the field will be considered invalid. */
+  /** Sets a custom validation message. Pass an empty string to restore validity. */
   setCustomValidity(message: string) {
     this.input.setCustomValidity(message);
-    this.invalid = !this.checkValidity();
+    this.formControlController.updateValidity();
   }
 
   render() {
@@ -428,7 +426,6 @@ export default class SlInput extends ShoelaceElement implements ShoelaceFormCont
               'input--disabled': this.disabled,
               'input--focused': this.hasFocus,
               'input--empty': !this.value,
-              'input--invalid': this.invalid,
               'input--no-spin-buttons': this.noSpinButtons,
               'input--is-firefox': isFirefox
             })}
@@ -460,7 +457,6 @@ export default class SlInput extends ShoelaceElement implements ShoelaceFormCont
               enterkeyhint=${ifDefined(this.enterkeyhint)}
               inputmode=${ifDefined(this.inputmode)}
               aria-describedby="help-text"
-              aria-invalid=${this.invalid ? 'true' : 'false'}
               @change=${this.handleChange}
               @input=${this.handleInput}
               @invalid=${this.handleInvalid}